Added Notes
Added notes, font for notes (Can not confirm we have full commercial permission to use it, but it sounds like we should...?) and fixed up the blueprints for it a bit.

Additionally, modified the map to make it easier for artists to use once they get into the engine.  (Change sun with up arrow's light level to change brightness, and remember to change skylight level back to 0.3 when done, and light source to 0.0001)