Basement Work + Inspection Light
All objects are now lit up as you inspect them, not just letters.  Additionally, added a basic puzzle to the basement, if you grab the -Lamp- you can use it to grab a key through the cell's door, or you can grab the key from the bloody bucket at the cost of some sanity.
If you have high sanity leaving the cell, there are no stairs.  But if you have low sanity, a set of very non-OSHA complaint stairs let you up to the next floor.