Added Footstep Sounds And AI Patroling
The AI now patrols between set patrol routes, simply add a patrol route and add the patrol points.
The AI will stop at each patrol point for a time before continuing, and it will teleport from the end of each route to the start of another randomly chosen route.
Because of the Teleportation, it's recommended to have the start and end of each patrol route outside the walls of the mansion, in the areas between rooms, or somewhere else out of sight from the player.